refactor(mesh): make category-editor media page only
